refactor: modularize authentication and password services, remove redundant code
- Extract authentication-specific logic into `OAuth2GoogleService` and `OAuth2Service` interfaces for better modularity.
- Introduce `PasswordService` and `PasswordController` to handle password-related operations, replacing `AuthService` functionality.
- Remove obsolete methods and tests related to `AuthService` and shift responsibilities to corresponding new services.
- Update configurations, controllers, and services to reflect these changes.

+++ src/main/java/com/nexaworks/rafiq/mapper/PatientMapper.java
@@ -14,7 +14,7 @@ public interface PatientMapper {
     @Mapping(target = "bmi", expression = "java(getBmiStatus(patient.getWeight(), patient.getHeight()))")
     PatientProfileResponse toResponse(Patient patient);
 
-    default String getBmiStatus(int weight, int height) {
+    default String getBmiStatus(double weight, int height) {
         int bmi = (int) (weight / (height * height));
         return bmi < 18.5 ? "Underweight" : bmi >= 18.5 && bmi < 25 ? "Normal" : "Overweight";
     }
